06_transport/transport-surface is updated to include dispersion effects
The older version did not test the dispersion since the value of parameter "alpha" was set to zero. Additionally, "alpha" represents dispersivity (m) which relevant to subsurface transport. This test is for surface transport, hence, "alpha" is replaced by "dispersion coefficient [m^2 s^-1]" with it's value equal to 1 m^2/s.

Another changes is to include "observation" list to record breakthrough curve at the end of the domain. This is just for demo to new users.