Finalized coupled_surface_pipe_model_bathymetry
This is similar to coupled_surface_pipe_model but with
bathymetry. Water initially flows from the surface down
in the drain reaching the pipe. The pipe then becomes
half pressurized. The manhole is in the pressurized part
so the flow of water reverses and goes from the pipe
back up to the surface.