Updates input files for lwc --> wc
Note these are not the same thing if molar and mass densities differ
by a different factor than the molecular mass of water (often the case
for constant density problems).  Probably we need to change the
surface-WC evaluator to deal with these cases correctly, or add
another surface-WC evaluator.

This doesn't update the gold yet.